refactor(web): improve separation of concerns in sports hall report feature
Split pdf-form-filler.ts (1158 lines) into focused modules:
- pdf-report-data.ts: Assignment data extraction, filename building
- pdf-field-mappings.ts: PDF field names, checklist sections, positions
- pdf-download.ts: Browser download trigger

Extract components and hooks for better modularity:
- NonConformantOverlay: Full-screen overlay from modal (reduces modal 673→170 lines)
- HappyPathContent, LanguageSelector, JerseyAdToggle, JerseyAdvertisingSection
- usePdfGeneration: Consolidates 3 duplicated PDF generation flows
- useBodyScrollLock, useOverlayTouchGuard: Shared overlay behavior hooks

Unify jersey advertising state at modal level (was duplicated in happy/NC paths).
Move coach names into useNonConformantWizard hook (fixes state sync issue).
Separate checklist UI model from PDF field mappings for cleaner layer boundaries.
