+++ packages/web/src/features/sports-hall-report/components/CommentStep.tsx
@@ -1,6 +1,6 @@
 import { AlertTriangle } from '@/shared/components/icons'
 import { useTranslation } from '@/shared/hooks/useTranslation'
-import type { ChecklistSection } from '@/shared/utils/pdf-form-filler'
+import type { ChecklistSection } from '@/shared/utils/pdf-field-mappings'
 
 interface CommentStepProps {
   sections: readonly ChecklistSection[]
