+++ packages/web/src/features/sports-hall-report/components/JerseyAdvertisingSection.tsx
@@ -1,0 +1,45 @@
+import { useTranslation } from '@/shared/hooks/useTranslation'
+import type { JerseyAdvertisingOptions } from '@/shared/utils/pdf-form-filler'
+
+import { JerseyAdToggle } from './JerseyAdToggle'
+
+interface JerseyAdvertisingSectionProps {
+  jerseyAdvertising: JerseyAdvertisingOptions
+  onToggleHome: () => void
+  onToggleAway: () => void
+  homeTeam: string
+  awayTeam: string
+  disabled: boolean
+}
+
+export function JerseyAdvertisingSection({
+  jerseyAdvertising,
+  onToggleHome,
+  onToggleAway,
+  homeTeam,
+  awayTeam,
+  disabled,
+}: JerseyAdvertisingSectionProps) {
+  const { tInterpolate } = useTranslation()
+
+  return (
+    <div className="space-y-1.5">
+      <JerseyAdToggle
+        label={tInterpolate('pdf.wizard.advertisingTeam', {
+          team: homeTeam || '–',
+        })}
+        checked={jerseyAdvertising.homeTeam}
+        onChange={onToggleHome}
+        disabled={disabled}
+      />
+      <JerseyAdToggle
+        label={tInterpolate('pdf.wizard.advertisingTeam', {
+          team: awayTeam || '–',
+        })}
+        checked={jerseyAdvertising.awayTeam}
+        onChange={onToggleAway}
+        disabled={disabled}
+      />
+    </div>
+  )
+}
