+++ packages/web/src/features/sports-hall-report/components/SignatureCanvas.tsx
@@ -5,6 +5,8 @@ import SignaturePad from 'signature_pad'
 
 import { Button } from '@/shared/components/Button'
 import { RotateCw, Trash2, Check, X } from '@/shared/components/icons'
+import { useBodyScrollLock } from '@/shared/hooks/useBodyScrollLock'
+import { useOverlayTouchGuard } from '@/shared/hooks/useOverlayTouchGuard'
 import { useTranslation } from '@/shared/hooks/useTranslation'
 
 /** Z-index above the modal overlay (50) so the signature canvas is on top */
@@ -30,6 +32,9 @@ export function SignatureCanvas({ onComplete, onCancel }: SignatureCanvasProps) 
   const [isPortrait, setIsPortrait] = useState(
     () => window.matchMedia('(orientation: portrait)').matches
   )
+
+  useBodyScrollLock(true)
+  const touchGuard = useOverlayTouchGuard()
 
   // Track orientation changes and enable/disable the pad accordingly
   useEffect(() => {
@@ -119,34 +124,6 @@ export function SignatureCanvas({ onComplete, onCancel }: SignatureCanvasProps) 
     onComplete(dataUrl)
   }, [onComplete])
 
-  // Lock body scroll and prevent swipe-to-navigate while overlay is open
-  useEffect(() => {
-    const prev = document.body.style.overflow
-    const prevOverscroll = document.body.style.overscrollBehavior
-    document.body.style.overflow = 'hidden'
-    document.body.style.overscrollBehavior = 'none'
-    return () => {
-      document.body.style.overflow = prev
-      document.body.style.overscrollBehavior = prevOverscroll
-    }
-  }, [])
-
-  // Stop all touch events from propagating through the React tree.
-  // Without this, React's synthetic events bubble from this portal through the
-  // React component hierarchy up to PullToRefresh, which interprets drawing
-  // strokes as pull-to-refresh gestures and unmounts the signature overlay.
-  const stopPropagation = useCallback((e: React.TouchEvent) => {
-    e.stopPropagation()
-  }, [])
-
-  // Prevent swipe-to-navigate gestures on the overlay container
-  const handleTouchMove = useCallback((e: React.TouchEvent) => {
-    e.stopPropagation()
-    // Allow touch events on the canvas (signature_pad handles them)
-    if (e.target instanceof HTMLCanvasElement) return
-    e.preventDefault()
-  }, [])
-
   // Portal to document.body so the overlay is outside the PullToRefresh DOM tree,
   // preventing pull-to-refresh gestures from interfering with signature drawing
   return createPortal(
@@ -156,9 +133,7 @@ export function SignatureCanvas({ onComplete, onCancel }: SignatureCanvasProps) 
       role="dialog"
       aria-modal="true"
       aria-label={t('pdf.wizard.signature.title')}
-      onTouchStart={stopPropagation}
-      onTouchMove={handleTouchMove}
-      onTouchEnd={stopPropagation}
+      {...touchGuard}
     >
       {/* Header toolbar */}
       <div className="flex items-center justify-between px-4 py-3 border-b border-gray-200 bg-gray-50">
