+++ packages/web/src/features/sports-hall-report/components/SubItemSelector.tsx
@@ -2,7 +2,8 @@ import { useCallback } from 'react'
 
 import { AlertTriangle, CheckCircle } from '@/shared/components/icons'
 import { useTranslation } from '@/shared/hooks/useTranslation'
-import type { ChecklistSection, NonConformantSelections } from '@/shared/utils/pdf-form-filler'
+import type { ChecklistSection } from '@/shared/utils/pdf-field-mappings'
+import type { NonConformantSelections } from '@/shared/utils/pdf-form-filler'
 
 interface SubItemSelectorProps {
   sections: readonly ChecklistSection[]
