+++ packages/web/src/features/sports-hall-report/utils/extractReportInfo.ts
@@ -1,12 +1,12 @@
 import type { Assignment } from '@/api/client'
-import type { LeagueCategory, SportsHallReportData } from '@/shared/utils/pdf-form-filler'
+import type { LeagueCategory, SportsHallReportData } from '@/shared/utils/pdf-report-data'
 
 export async function extractReportInfo(assignment: Assignment): Promise<{
   reportData: SportsHallReportData
   leagueCategory: LeagueCategory
 } | null> {
   const { extractSportsHallReportData, getLeagueCategoryFromAssignment } =
-    await import('@/shared/utils/pdf-form-filler')
+    await import('@/shared/utils/pdf-report-data')
 
   const reportData = extractSportsHallReportData(assignment)
   const leagueCategory = getLeagueCategoryFromAssignment(assignment)
