+++ packages/web/src/shared/utils/pdf-download.ts
@@ -1,0 +1,11 @@
+export function downloadPdf(pdfBytes: Uint8Array, filename: string): void {
+  const blob = new Blob([new Uint8Array(pdfBytes)], { type: 'application/pdf' })
+  const url = URL.createObjectURL(blob)
+  const link = document.createElement('a')
+  link.href = url
+  link.download = filename
+  document.body.appendChild(link)
+  link.click()
+  document.body.removeChild(link)
+  URL.revokeObjectURL(url)
+}
