+++ packages/web/src/shared/utils/pdf-form-filler.ts
@@ -1,146 +1,81 @@
-import { format } from 'date-fns'
+/**
+ * PDF form filler — core PDF manipulation logic.
+ *
+ * Re-exports types and functions from split modules for backward compatibility.
+ * New code should import directly from the specific module:
+ *   - pdf-report-data.ts  — Assignment data extraction, filename building
+ *   - pdf-field-mappings.ts — PDF field names, checklist sections, positions
+ *   - pdf-download.ts — Browser download trigger
+ */
 
-import type { Assignment } from '@/api/client'
-import type { TranslationKey } from '@/i18n'
 import { logger } from '@/shared/utils/logger'
 
-export type LeagueCategory = 'NLA' | 'NLB'
-export type Language = 'de' | 'fr'
-export type Gender = 'm' | 'f'
+import { downloadPdf } from './pdf-download'
+import {
+  ALL_SIGNATURE_POSITIONS,
+  COMMENT_FIELD_FONT_SIZE,
+  RADIO_NOT_OK_OPTION,
+  RADIO_OK_OPTION,
+  SIGNATURE_POSITIONS,
+  getFieldMapping,
+  getPdfChecklistSections,
+  getSignatureNameFields,
+  getWizardFieldMapping,
+  type AllSignaturePositions,
+  type PdfChecklistSection,
+  type SignaturePosition,
+} from './pdf-field-mappings'
+import {
+  buildReportFilename,
+  type Language,
+  type LeagueCategory,
+  type SportsHallReportData,
+} from './pdf-report-data'
 
-export function mapAppLocaleToPdfLanguage(appLocale: string): Language {
-  if (appLocale === 'fr' || appLocale === 'it') {
-    return 'fr'
-  }
-  return 'de'
+// ---------------------------------------------------------------------------
+// Re-exports for backward compatibility
+// ---------------------------------------------------------------------------
+
+export { downloadPdf } from './pdf-download'
+export {
+  getChecklistSections,
+  type ChecklistSection,
+  type ChecklistSubItem,
+} from './pdf-field-mappings'
+export {
+  buildReportFilename,
+  extractSportsHallReportData,
+  getLeagueCategoryFromAssignment,
+  mapAppLocaleToPdfLanguage,
+  type Gender,
+  type Language,
+  type LeagueCategory,
+  type SportsHallReportData,
+} from './pdf-report-data'
+
+// ---------------------------------------------------------------------------
+// Shared types
+// ---------------------------------------------------------------------------
+
+export interface JerseyAdvertisingOptions {
+  homeTeam: boolean
+  awayTeam: boolean
 }
 
-export interface SportsHallReportData {
-  gameNumber: string
-  homeTeam: string
-  awayTeam: string
-  gender: Gender
-  hallName: string
-  location: string
-  date: string
-  startingDateTime?: string
-  firstRefereeName?: string
-  secondRefereeName?: string
+/** Which sub-items are flagged as not-OK, keyed by section ID then sub-item ID. */
+export type NonConformantSelections = Record<string, Set<string>>
+
+/** Signature data for the non-conformant workflow (up to 4 signers). */
+export interface NonConformantSignatures {
+  firstReferee?: string
+  secondReferee?: string
+  homeTeamCoach?: { name: string; signature: string }
+  awayTeamCoach?: { name: string; signature: string }
 }
 
-function formatDateForReport(isoString: string | undefined): string {
-  if (!isoString) return ''
-  try {
-    return format(new Date(isoString), 'dd.MM.yy')
-  } catch {
-    // Invalid date format - return empty string so form field shows blank
-    logger.warn('Failed to parse date for PDF report:', isoString)
-    return ''
-  }
-}
-
-function getRefereeName(
-  convocation:
-    | {
-        indoorAssociationReferee?: {
-          indoorReferee?: {
-            person?: {
-              firstName?: string
-              lastName?: string
-              displayName?: string
-            }
-          }
-        }
-      }
-    | null
-    | undefined
-): string | undefined {
-  const person = convocation?.indoorAssociationReferee?.indoorReferee?.person
-  if (!person) return undefined
-  return (
-    person.displayName || `${person.firstName ?? ''} ${person.lastName ?? ''}`.trim() || undefined
-  )
-}
-
-export function extractSportsHallReportData(assignment: Assignment): SportsHallReportData | null {
-  const game = assignment.refereeGame?.game
-  if (!game) return null
-
-  const leagueCategoryName = game.group?.phase?.league?.leagueCategory?.name
-  if (leagueCategoryName !== 'NLA' && leagueCategoryName !== 'NLB') {
-    return null
-  }
-
-  const refereeGame = assignment.refereeGame
-  const firstReferee = getRefereeName(refereeGame?.activeRefereeConvocationFirstHeadReferee)
-  const secondReferee = getRefereeName(refereeGame?.activeRefereeConvocationSecondHeadReferee)
-
-  const reportData: SportsHallReportData = {
-    gameNumber: game.number?.toString() ?? '',
-    homeTeam: game.encounter?.teamHome?.name ?? '',
-    awayTeam: game.encounter?.teamAway?.name ?? '',
-    gender: (game.group?.phase?.league?.gender ?? 'm') as Gender,
-    hallName: game.hall?.name ?? '',
-    location: game.hall?.primaryPostalAddress?.city ?? '',
-    date: formatDateForReport(game.startingDateTime),
-    startingDateTime: game.startingDateTime,
-    firstRefereeName: firstReferee,
-    secondRefereeName: secondReferee,
-  }
-
-  // Debug logging for troubleshooting PDF generation
-  logger.debug('[pdf-form-filler] Extracted report data:', reportData)
-
-  return reportData
-}
-
-export function getLeagueCategoryFromAssignment(assignment: Assignment): LeagueCategory | null {
-  const name = assignment.refereeGame?.game?.group?.phase?.league?.leagueCategory?.name
-  if (name === 'NLA' || name === 'NLB') {
-    return name
-  }
-  return null
-}
-
-interface FieldMapping {
-  gameNumber: string
-  homeTeam: string
-  awayTeam: string
-  genderRadio: string
-  hallName: string
-  location: string
-  date: string
-  firstRefereeName: string
-  secondRefereeName: string
-}
-
-const NLA_FIELD_MAPPING: FieldMapping = {
-  gameNumber: 'SpielNr',
-  homeTeam: 'Heimteam',
-  awayTeam: 'Gastteam',
-  genderRadio: 'Gruppe3',
-  hallName: 'Hallenname',
-  location: 'Ort',
-  date: 'Datum',
-  firstRefereeName: 'Text19',
-  secondRefereeName: 'Text20',
-}
-
-const NLB_FIELD_MAPPING: FieldMapping = {
-  gameNumber: 'Text9',
-  homeTeam: 'Text10',
-  awayTeam: 'Text12',
-  genderRadio: 'Gruppe15',
-  hallName: 'Text14',
-  location: 'Text13',
-  date: 'Text11',
-  firstRefereeName: 'Text23',
-  secondRefereeName: 'Text24',
-}
-
-function getFieldMapping(leagueCategory: LeagueCategory): FieldMapping {
-  return leagueCategory === 'NLA' ? NLA_FIELD_MAPPING : NLB_FIELD_MAPPING
-}
+// ---------------------------------------------------------------------------
+// Internal helpers
+// ---------------------------------------------------------------------------
 
 /**
  * Safely sets a text field in a PDF form, logging a warning on failure.
@@ -158,10 +93,9 @@ function trySetTextField(form: any, fieldName: string, value: string | undefined
 
 /**
  * Fills the basic game info fields (teams, date, hall, referees, gender) in a PDF form.
- * Shared between the standard report and the wizard report.
  */
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-function fillBaseGameInfo(form: any, data: SportsHallReportData, mapping: FieldMapping): void {
+function fillBaseGameInfo(form: any, data: SportsHallReportData, mapping: ReturnType<typeof getFieldMapping>): void {
   trySetTextField(form, mapping.gameNumber, data.gameNumber)
   trySetTextField(form, mapping.homeTeam, data.homeTeam)
   trySetTextField(form, mapping.awayTeam, data.awayTeam)
@@ -221,539 +155,15 @@ async function loadPdfForm(
   return { pdfDoc, form }
 }
 
-export async function fillSportsHallReportForm(
-  data: SportsHallReportData,
-  leagueCategory: LeagueCategory,
-  language: Language
-): Promise<Uint8Array> {
-  const { pdfDoc, form } = await loadPdfForm(leagueCategory, language)
-  const mapping = getFieldMapping(leagueCategory)
-
-  fillBaseGameInfo(form, data, mapping)
-
-  // Do not flatten: referees download this form to fill remaining fields themselves
-  return pdfDoc.save()
-}
-
-export function downloadPdf(pdfBytes: Uint8Array, filename: string): void {
-  const blob = new Blob([new Uint8Array(pdfBytes)], { type: 'application/pdf' })
-  const url = URL.createObjectURL(blob)
-  const link = document.createElement('a')
-  link.href = url
-  link.download = filename
-  document.body.appendChild(link)
-  link.click()
-  document.body.removeChild(link)
-  URL.revokeObjectURL(url)
-}
-
-const REPORT_NAME: Record<Language, string> = {
-  de: 'hallenrapport',
-  fr: 'rapport_salle',
-}
-
-export function buildReportFilename(
-  leagueCategory: LeagueCategory,
-  language: Language,
-  startingDateTime?: string,
-  gameNumber?: string
-): string {
-  const league = leagueCategory.toLowerCase()
-  const reportName = REPORT_NAME[language]
-  let datePart = 'unknown'
-  if (startingDateTime) {
-    try {
-      datePart = format(new Date(startingDateTime), 'yyyyMMdd')
-    } catch {
-      logger.warn('Failed to parse date for report filename:', startingDateTime)
-    }
-  }
-  const suffix = gameNumber ? `_${gameNumber}` : ''
-  return `${league}_${reportName}_${datePart}${suffix}.pdf`
-}
-
-export async function generateAndDownloadSportsHallReport(
-  data: SportsHallReportData,
-  leagueCategory: LeagueCategory,
-  language: Language
-): Promise<void> {
-  const pdfBytes = await fillSportsHallReportForm(data, leagueCategory, language)
-  const filename = buildReportFilename(
-    leagueCategory,
-    language,
-    data.startingDateTime,
-    data.gameNumber
-  )
-  downloadPdf(pdfBytes, filename)
-}
-
-/**
- * Wizard-specific field mapping for checkbox/radio fields that the wizard fills
- * in addition to the base game info fields.
- */
-interface WizardFieldMapping {
-  /** Checkbox field name for "Tous les points sont en ordre" / "Alle Punkte in Ordnung" */
-  allPointsInOrderCheckbox: string
-  /** Radio group field name for home team "Werbung auf Spielerkleidung" (Ja/Nein) */
-  advertisingHomeTeam: string
-  /** Radio group field name for away team "Werbung auf Spielerkleidung" (Ja/Nein) */
-  advertisingAwayTeam: string
-}
-
-const NLA_WIZARD_FIELDS: WizardFieldMapping = {
-  allPointsInOrderCheckbox: 'Kontrollkästchen8',
-  advertisingHomeTeam: 'Gruppe430',
-  advertisingAwayTeam: 'Gruppe434',
-}
-
-const NLB_WIZARD_FIELDS: WizardFieldMapping = {
-  allPointsInOrderCheckbox: 'Kontrollkästchen17',
-  advertisingHomeTeam: '31',
-  advertisingAwayTeam: '34',
-}
-
-/** OK option value used by all radio groups in the PDF templates */
-const RADIO_OK_OPTION = 'Auswahl3'
-
-/** Not-OK option value used by all radio groups in the PDF templates */
-const RADIO_NOT_OK_OPTION = 'Auswahl4'
-
-export interface JerseyAdvertisingOptions {
-  homeTeam: boolean
-  awayTeam: boolean
-}
-
 // ---------------------------------------------------------------------------
-// Checklist section / sub-item mappings for the non-conformant workflow
+// Signature embedding
 // ---------------------------------------------------------------------------
 
-export interface ChecklistSubItem {
-  /** Unique key within the section (e.g. 'lines', 'attackLine') */
-  id: string
-  /** i18n key for the sub-item label */
-  labelKey: TranslationKey
-  /** PDF radio group field name(s) — one per sub-item row on the form */
-  radioField: string
-}
+const SIGNATURE_CROP_PADDING = 10
+const SIGNATURE_EMBED_PADDING_PT = 3
+const RGBA_CHANNELS = 4
+const ALPHA_CHANNEL_OFFSET = 3
 
-export interface ChecklistSection {
-  /** Section letter matching the PDF form (A, B, C, …) */
-  id: string
-  /** i18n key for the section label */
-  labelKey: TranslationKey
-  /** Sub-items within this section */
-  subItems: readonly ChecklistSubItem[]
-  /** PDF text field name(s) for the remarks column */
-  commentFields: readonly string[]
-}
-
-const NLA_CHECKLIST_SECTIONS: readonly ChecklistSection[] = [
-  {
-    id: 'A',
-    labelKey: 'pdf.wizard.sections.fieldAndLines',
-    subItems: [
-      { id: 'lines', labelKey: 'pdf.wizard.subItems.boundaryLines', radioField: 'Gruppe4' },
-      { id: 'attackLine', labelKey: 'pdf.wizard.subItems.attackLine', radioField: 'Gruppe41' },
-    ],
-    commentFields: ['Text7.0.0'],
-  },
-  {
-    id: 'B',
-    labelKey: 'pdf.wizard.sections.warmUpArea',
-    subItems: [
-      { id: 'warmUp', labelKey: 'pdf.wizard.subItems.warmUpArea', radioField: 'Gruppe42' },
-    ],
-    commentFields: ['Text7.0.1'],
-  },
-  {
-    id: 'C',
-    labelKey: 'pdf.wizard.sections.netEquipment',
-    subItems: [
-      { id: 'antennas', labelKey: 'pdf.wizard.subItems.antennas', radioField: 'Gruppe43' },
-      { id: 'netBand', labelKey: 'pdf.wizard.subItems.netTopBand', radioField: 'Gruppe44' },
-      { id: 'reserveNet', labelKey: 'pdf.wizard.subItems.reserveNet', radioField: 'Gruppe45' },
-      { id: 'postsNoWire', labelKey: 'pdf.wizard.subItems.postsNoWire', radioField: 'Gruppe46' },
-      { id: 'postsPadding', labelKey: 'pdf.wizard.subItems.postsPadding', radioField: 'Gruppe47' },
-      { id: 'tabletMount', labelKey: 'pdf.wizard.subItems.tabletMount', radioField: 'Gruppe48' },
-      { id: 'netHeight', labelKey: 'pdf.wizard.subItems.netHeight', radioField: 'Gruppe49' },
-    ],
-    commentFields: ['Text7.0.2'],
-  },
-  {
-    id: 'D',
-    labelKey: 'pdf.wizard.sections.refereeChair',
-    subItems: [{ id: 'chair', labelKey: 'pdf.wizard.subItems.refereeChair', radioField: '10' }],
-    commentFields: ['Text7.0.3'],
-  },
-  {
-    id: 'E',
-    labelKey: 'pdf.wizard.sections.manometer',
-    subItems: [{ id: 'manometer', labelKey: 'pdf.wizard.subItems.manometer', radioField: '11' }],
-    commentFields: ['Text7.0.4'],
-  },
-  {
-    id: 'F',
-    labelKey: 'pdf.wizard.sections.measuringRod',
-    subItems: [{ id: 'rod', labelKey: 'pdf.wizard.subItems.measuringRod', radioField: '12' }],
-    commentFields: ['Text7.0.5'],
-  },
-  {
-    id: 'G',
-    labelKey: 'pdf.wizard.sections.lineFlags',
-    subItems: [{ id: 'flags', labelKey: 'pdf.wizard.subItems.lineFlags', radioField: '13' }],
-    commentFields: ['Text7.0.6'],
-  },
-  {
-    id: 'H',
-    labelKey: 'pdf.wizard.sections.buzzer',
-    subItems: [{ id: 'buzzer', labelKey: 'pdf.wizard.subItems.buzzer', radioField: '14' }],
-    commentFields: ['Text7.0.7'],
-  },
-  {
-    id: 'I',
-    labelKey: 'pdf.wizard.sections.eScoresheet',
-    subItems: [
-      { id: 'eScorerOnTime', labelKey: 'pdf.wizard.subItems.eScorerOnTime', radioField: '15' },
-      { id: 'reserveLaptop', labelKey: 'pdf.wizard.subItems.reserveLaptop', radioField: '16' },
-      { id: 'usbStick', labelKey: 'pdf.wizard.subItems.usbStick', radioField: '17' },
-      {
-        id: 'reserveMatchSheet',
-        labelKey: 'pdf.wizard.subItems.reserveMatchSheet',
-        radioField: '18',
-      },
-    ],
-    commentFields: ['Text7.1'],
-  },
-  {
-    id: 'J',
-    labelKey: 'pdf.wizard.sections.tablets',
-    subItems: [
-      { id: 'refereeTablets', labelKey: 'pdf.wizard.subItems.refereeTablets', radioField: '19' },
-      { id: 'benchTablets', labelKey: 'pdf.wizard.subItems.benchTablets', radioField: '20' },
-    ],
-    commentFields: ['Text7.2'],
-  },
-  {
-    id: 'K',
-    labelKey: 'pdf.wizard.sections.scoreboard',
-    subItems: [
-      { id: 'scoreboard', labelKey: 'pdf.wizard.subItems.scoreboard', radioField: '21' },
-      {
-        id: 'scoreboardFunction',
-        labelKey: 'pdf.wizard.subItems.scoreboardFunction',
-        radioField: '22',
-      },
-    ],
-    commentFields: ['Text7.3'],
-  },
-  {
-    id: 'L',
-    labelKey: 'pdf.wizard.sections.balls',
-    subItems: [
-      { id: 'ballCount', labelKey: 'pdf.wizard.subItems.ballCount', radioField: '23' },
-      {
-        id: 'ballCondition',
-        labelKey: 'pdf.wizard.subItems.ballCondition',
-        radioField: 'Gruppe424',
-      },
-    ],
-    commentFields: ['Text7.4'],
-  },
-  {
-    id: 'M',
-    labelKey: 'pdf.wizard.sections.ballRetrievers',
-    subItems: [
-      {
-        id: 'ballRetrievers',
-        labelKey: 'pdf.wizard.subItems.ballRetrievers',
-        radioField: 'Gruppe425',
-      },
-    ],
-    commentFields: ['Text7.5'],
-  },
-  {
-    id: 'N',
-    labelKey: 'pdf.wizard.sections.quickMoppers',
-    subItems: [
-      { id: 'quickMoppers', labelKey: 'pdf.wizard.subItems.quickMoppers', radioField: 'Gruppe426' },
-    ],
-    commentFields: ['Text7.6'],
-  },
-  {
-    id: 'O',
-    labelKey: 'pdf.wizard.sections.hallSpeaker',
-    subItems: [
-      { id: 'hallSpeaker', labelKey: 'pdf.wizard.subItems.hallSpeaker', radioField: 'Gruppe427' },
-    ],
-    commentFields: ['Text7.7'],
-  },
-  {
-    id: 'P',
-    labelKey: 'pdf.wizard.sections.dressHome',
-    subItems: [
-      { id: 'dressHomeColor', labelKey: 'pdf.wizard.subItems.dressColor', radioField: 'Gruppe428' },
-      {
-        id: 'dressHomeLibero',
-        labelKey: 'pdf.wizard.subItems.liberoDress',
-        radioField: 'Gruppe429',
-      },
-      {
-        id: 'dressHomeAd',
-        labelKey: 'pdf.wizard.subItems.advertisingOnUniform',
-        radioField: 'Gruppe430',
-      },
-    ],
-    commentFields: ['Text7.8'],
-  },
-  {
-    id: 'Q',
-    labelKey: 'pdf.wizard.sections.dressAway',
-    subItems: [
-      { id: 'dressAwayColor', labelKey: 'pdf.wizard.subItems.dressColor', radioField: 'Gruppe431' },
-      {
-        id: 'dressAwayLibero',
-        labelKey: 'pdf.wizard.subItems.liberoDress',
-        radioField: 'Gruppe432',
-      },
-      {
-        id: 'dressAwayMatchkit',
-        labelKey: 'pdf.wizard.subItems.matchKit',
-        radioField: 'Gruppe433',
-      },
-      {
-        id: 'dressAwayAd',
-        labelKey: 'pdf.wizard.subItems.advertisingOnUniform',
-        radioField: 'Gruppe434',
-      },
-    ],
-    commentFields: ['Text7.9'],
-  },
-  {
-    id: 'R',
-    labelKey: 'pdf.wizard.sections.miscellaneous',
-    subItems: [
-      { id: 'misc1', labelKey: 'pdf.wizard.subItems.miscItem1', radioField: 'Gruppe435' },
-      { id: 'misc2', labelKey: 'pdf.wizard.subItems.miscItem2', radioField: 'Gruppe440' },
-    ],
-    commentFields: ['Text7.10', 'Text7.11'],
-  },
-] as const
-
-const NLB_CHECKLIST_SECTIONS: readonly ChecklistSection[] = [
-  {
-    id: 'A',
-    labelKey: 'pdf.wizard.sections.fieldAndLines',
-    subItems: [
-      { id: 'lines', labelKey: 'pdf.wizard.subItems.boundaryLines', radioField: 'Gruppe16' },
-      { id: 'attackLine', labelKey: 'pdf.wizard.subItems.attackLine', radioField: 'Gruppe17' },
-    ],
-    commentFields: ['Text16.0.0'],
-  },
-  {
-    id: 'B',
-    labelKey: 'pdf.wizard.sections.netEquipment',
-    subItems: [
-      { id: 'antennas', labelKey: 'pdf.wizard.subItems.antennas', radioField: 'Gruppe18' },
-      { id: 'netBand', labelKey: 'pdf.wizard.subItems.netTopBand', radioField: 'Gruppe19' },
-      {
-        id: 'netBottomBand',
-        labelKey: 'pdf.wizard.subItems.netBottomBand',
-        radioField: 'Gruppe20',
-      },
-    ],
-    commentFields: ['Text16.0.1'],
-  },
-  {
-    id: 'C',
-    labelKey: 'pdf.wizard.sections.numberPlates',
-    subItems: [
-      {
-        id: 'numberPlates',
-        labelKey: 'pdf.wizard.subItems.numberPlates',
-        radioField: 'Gruppe21',
-      },
-    ],
-    commentFields: ['Text16.1'],
-  },
-  {
-    id: 'D',
-    labelKey: 'pdf.wizard.sections.manometer',
-    subItems: [{ id: 'manometer', labelKey: 'pdf.wizard.subItems.manometer', radioField: '22' }],
-    commentFields: ['Text16.2'],
-  },
-  {
-    id: 'E',
-    labelKey: 'pdf.wizard.sections.measuringRod',
-    subItems: [{ id: 'rod', labelKey: 'pdf.wizard.subItems.measuringRod', radioField: '23' }],
-    commentFields: ['Text16.3'],
-  },
-  {
-    id: 'F',
-    labelKey: 'pdf.wizard.sections.balls',
-    subItems: [
-      {
-        id: 'practiseBalls',
-        labelKey: 'pdf.wizard.subItems.practiseBalls',
-        radioField: '24',
-      },
-      { id: 'matchBalls', labelKey: 'pdf.wizard.subItems.matchBalls', radioField: '25' },
-      {
-        id: 'matchBallsException',
-        labelKey: 'pdf.wizard.subItems.matchBallsException',
-        radioField: '26',
-      },
-    ],
-    commentFields: ['Text16.4'],
-  },
-  {
-    id: 'G',
-    labelKey: 'pdf.wizard.sections.ballRetrievers',
-    subItems: [
-      { id: 'ballRetrievers', labelKey: 'pdf.wizard.subItems.ballRetrievers', radioField: '27' },
-      {
-        id: 'ballRetrieversException',
-        labelKey: 'pdf.wizard.subItems.ballRetrieversException',
-        radioField: '28',
-      },
-    ],
-    commentFields: ['Text16.5'],
-  },
-  {
-    id: 'H',
-    labelKey: 'pdf.wizard.sections.dressHome',
-    subItems: [
-      { id: 'dressHomeColor', labelKey: 'pdf.wizard.subItems.dressColor', radioField: '29' },
-      { id: 'dressHomeLibero', labelKey: 'pdf.wizard.subItems.liberoDress', radioField: '30' },
-      {
-        id: 'dressHomeAd',
-        labelKey: 'pdf.wizard.subItems.advertisingOnUniform',
-        radioField: '31',
-      },
-    ],
-    commentFields: ['Text16.6'],
-  },
-  {
-    id: 'I',
-    labelKey: 'pdf.wizard.sections.dressAway',
-    subItems: [
-      { id: 'dressAwayColor', labelKey: 'pdf.wizard.subItems.dressColor', radioField: '32' },
-      { id: 'dressAwayLibero', labelKey: 'pdf.wizard.subItems.liberoDress', radioField: '33' },
-      {
-        id: 'dressAwayAd',
-        labelKey: 'pdf.wizard.subItems.advertisingOnUniform',
-        radioField: '34',
-      },
-    ],
-    commentFields: ['Text16.7'],
-  },
-  {
-    id: 'J',
-    labelKey: 'pdf.wizard.sections.eScoresheet',
-    subItems: [
-      { id: 'eScorerOnTime', labelKey: 'pdf.wizard.subItems.eScorerOnTime', radioField: '35' },
-      { id: 'reserveLaptop', labelKey: 'pdf.wizard.subItems.reserveLaptop', radioField: '36' },
-      { id: 'usbStick', labelKey: 'pdf.wizard.subItems.usbStick', radioField: '37' },
-      {
-        id: 'reserveMatchSheet',
-        labelKey: 'pdf.wizard.subItems.reserveMatchSheet',
-        radioField: '38',
-      },
-    ],
-    commentFields: ['Text16.8'],
-  },
-  {
-    id: 'K',
-    labelKey: 'pdf.wizard.sections.tablets',
-    subItems: [
-      { id: 'refereeTablets', labelKey: 'pdf.wizard.subItems.refereeTablets', radioField: '39' },
-    ],
-    commentFields: ['Text16.9'],
-  },
-  {
-    id: 'L',
-    labelKey: 'pdf.wizard.sections.miscellaneous',
-    subItems: [
-      { id: 'misc1', labelKey: 'pdf.wizard.subItems.miscItem1', radioField: '40' },
-      { id: 'misc2', labelKey: 'pdf.wizard.subItems.miscItem2', radioField: '41' },
-    ],
-    commentFields: ['Text16.10', 'Text16.11'],
-  },
-] as const
-
-export function getChecklistSections(leagueCategory: LeagueCategory): readonly ChecklistSection[] {
-  return leagueCategory === 'NLA' ? NLA_CHECKLIST_SECTIONS : NLB_CHECKLIST_SECTIONS
-}
-
-/**
- * All four signature name text fields at the bottom of the report form.
- */
-interface SignatureNameFields {
-  firstReferee: string
-  secondReferee: string
-  homeTeam: string
-  awayTeam: string
-}
-
-const NLA_SIGNATURE_NAME_FIELDS: SignatureNameFields = {
-  firstReferee: 'Text19',
-  secondReferee: 'Text20',
-  homeTeam: 'Text21',
-  awayTeam: 'Text22',
-}
-
-const NLB_SIGNATURE_NAME_FIELDS: SignatureNameFields = {
-  firstReferee: 'Text23',
-  secondReferee: 'Text24',
-  homeTeam: 'Text25',
-  awayTeam: 'Text26',
-}
-
-function getWizardFieldMapping(leagueCategory: LeagueCategory): WizardFieldMapping {
-  return leagueCategory === 'NLA' ? NLA_WIZARD_FIELDS : NLB_WIZARD_FIELDS
-}
-
-/**
- * Signature placement coordinates per league category.
- * Positioned to the right of the referee name text fields at the bottom of the page.
- */
-interface SignaturePosition {
-  x: number
-  y: number
-  width: number
-  height: number
-}
-
-const SIGNATURE_POSITIONS: Record<LeagueCategory, SignaturePosition> = {
-  NLA: { x: 340, y: 104, width: 130, height: 24 },
-  NLB: { x: 340, y: 157, width: 130, height: 24 },
-}
-
-/** Signature positions for all four signers (non-conformant workflow). */
-interface AllSignaturePositions {
-  firstReferee: SignaturePosition
-  secondReferee: SignaturePosition
-  homeTeam: SignaturePosition
-  awayTeam: SignaturePosition
-}
-
-const ALL_SIGNATURE_POSITIONS: Record<LeagueCategory, AllSignaturePositions> = {
-  NLA: {
-    firstReferee: { x: 340, y: 104, width: 130, height: 24 },
-    secondReferee: { x: 340, y: 83, width: 130, height: 18 },
-    homeTeam: { x: 340, y: 62, width: 130, height: 18 },
-    awayTeam: { x: 340, y: 41, width: 130, height: 18 },
-  },
-  NLB: {
-    firstReferee: { x: 340, y: 157, width: 130, height: 24 },
-    secondReferee: { x: 340, y: 129, width: 130, height: 18 },
-    homeTeam: { x: 340, y: 101, width: 130, height: 18 },
-    awayTeam: { x: 340, y: 73, width: 130, height: 18 },
-  },
-}
-
-/**
- * Decode a data URL (e.g. from canvas.toDataURL) to raw bytes.
- */
 function dataUrlToBytes(dataUrl: string): Uint8Array {
   const base64 = dataUrl.split(',')[1]
   if (!base64) throw new Error('Invalid data URL')
@@ -765,15 +175,6 @@ function dataUrlToBytes(dataUrl: string): Uint8Array {
   return bytes
 }
 
-const SIGNATURE_CROP_PADDING = 10
-const SIGNATURE_EMBED_PADDING_PT = 3
-const RGBA_CHANNELS = 4
-const ALPHA_CHANNEL_OFFSET = 3
-
-/**
- * Crop a signature PNG data URL to its bounding box, removing transparent
- * whitespace around the actual strokes. Returns the cropped data URL.
- */
 async function cropSignatureDataUrl(dataUrl: string): Promise<string> {
   const img = new Image()
   await new Promise<void>((resolve, reject) => {
@@ -792,7 +193,6 @@ async function cropSignatureDataUrl(dataUrl: string): Promise<string> {
   const imageData = ctx.getImageData(0, 0, canvas.width, canvas.height)
   const { data, width, height } = imageData
 
-  // Find bounding box of non-transparent pixels
   let minX = width
   let minY = height
   let maxX = 0
@@ -810,10 +210,8 @@ async function cropSignatureDataUrl(dataUrl: string): Promise<string> {
     }
   }
 
-  // No visible pixels found — return original
   if (maxX < minX || maxY < minY) return dataUrl
 
-  // Add padding, clamped to canvas bounds
   const cropX = Math.max(0, minX - SIGNATURE_CROP_PADDING)
   const cropY = Math.max(0, minY - SIGNATURE_CROP_PADDING)
   const cropW = Math.min(width - cropX, maxX - minX + 1 + SIGNATURE_CROP_PADDING * 2)
@@ -829,9 +227,6 @@ async function cropSignatureDataUrl(dataUrl: string): Promise<string> {
   return croppedCanvas.toDataURL('image/png')
 }
 
-/**
- * Embed a PNG signature image into a PDF page at the given position.
- */
 async function embedSignatureAtPosition(
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   pdfDoc: any,
@@ -844,7 +239,6 @@ async function embedSignatureAtPosition(
 
   const page = pdfDoc.getPage(0)
 
-  // Scale the signature to fit within the field with padding
   const padding = SIGNATURE_EMBED_PADDING_PT
   const maxWidth = pos.width - padding * 2
   const maxHeight = pos.height - padding * 2
@@ -856,7 +250,6 @@ async function embedSignatureAtPosition(
     drawWidth = drawHeight * aspectRatio
   }
 
-  // Center the signature within the field
   const xOffset = pos.x + padding + (maxWidth - drawWidth) / 2
   const yOffset = pos.y + padding + (maxHeight - drawHeight) / 2
 
@@ -868,10 +261,6 @@ async function embedSignatureAtPosition(
   })
 }
 
-/**
- * Embed a PNG signature image into a PDF document at the first referee's
- * signature position.
- */
 async function embedSignature(
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   pdfDoc: any,
@@ -881,15 +270,132 @@ async function embedSignature(
   await embedSignatureAtPosition(pdfDoc, signatureDataUrl, SIGNATURE_POSITIONS[leagueCategory])
 }
 
-/**
- * Fills the sports hall report with "all points in order" checked and
- * advertising declared as "Ja" for both teams. Individual checklist items
- * are left unchecked — the referee only needs to fill these if something
- * is not in order. This is the "happy path" wizard flow.
- *
- * When a signatureDataUrl is provided, the first referee's signature is
- * embedded at the designated position on the PDF.
- */
+async function embedAllSignatures(
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  pdfDoc: any,
+  positions: AllSignaturePositions,
+  signatures: NonConformantSignatures
+): Promise<void> {
+  const entries: Array<[string | undefined, SignaturePosition]> = [
+    [signatures.firstReferee, positions.firstReferee],
+    [signatures.secondReferee, positions.secondReferee],
+    [signatures.homeTeamCoach?.signature, positions.homeTeam],
+    [signatures.awayTeamCoach?.signature, positions.awayTeam],
+  ]
+
+  for (const [dataUrl, position] of entries) {
+    if (!dataUrl) continue
+    try {
+      await embedSignatureAtPosition(pdfDoc, dataUrl, position)
+    } catch (error) {
+      logger.warn('Could not embed signature:', error)
+    }
+  }
+}
+
+// ---------------------------------------------------------------------------
+// Checklist radio / comment filling
+// ---------------------------------------------------------------------------
+
+function fillChecklistRadios(
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  form: any,
+  sections: readonly PdfChecklistSection[],
+  nonConformantSubItems: NonConformantSelections
+): void {
+  for (const section of sections) {
+    const flaggedItems = nonConformantSubItems[section.id]
+    for (const subItem of section.subItems) {
+      const isNotOk = flaggedItems?.has(subItem.id) ?? false
+      const option = isNotOk ? RADIO_NOT_OK_OPTION : RADIO_OK_OPTION
+      try {
+        form.getRadioGroup(subItem.radioField).select(option)
+      } catch (error) {
+        logger.warn(`Could not set radio "${subItem.radioField}" to ${option}:`, error)
+      }
+    }
+  }
+}
+
+function fillNonConformantComments(
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  form: any,
+  sections: readonly PdfChecklistSection[],
+  nonConformantSubItems: NonConformantSelections,
+  sectionComments: Record<string, string>
+): void {
+  for (const section of sections) {
+    if (!nonConformantSubItems[section.id]?.size) continue
+    const comment = sectionComments[section.id]?.trim()
+    if (!comment) continue
+    const commentField = section.commentFields[0]
+    if (!commentField) continue
+    try {
+      const field = form.getTextField(commentField)
+      field.setFontSize(COMMENT_FIELD_FONT_SIZE)
+      field.setText(comment)
+    } catch (error) {
+      logger.warn(`Could not set comment field "${commentField}":`, error)
+    }
+  }
+}
+
+// ---------------------------------------------------------------------------
+// Advertising radio helper
+// ---------------------------------------------------------------------------
+
+function fillAdvertisingRadios(
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  form: any,
+  wizardMapping: ReturnType<typeof getWizardFieldMapping>,
+  jerseyAdvertising: JerseyAdvertisingOptions
+): void {
+  const adEntries: Array<{ field: string; hasAd: boolean }> = [
+    { field: wizardMapping.advertisingHomeTeam, hasAd: jerseyAdvertising.homeTeam },
+    { field: wizardMapping.advertisingAwayTeam, hasAd: jerseyAdvertising.awayTeam },
+  ]
+  for (const { field, hasAd } of adEntries) {
+    try {
+      form.getRadioGroup(field).select(hasAd ? RADIO_OK_OPTION : RADIO_NOT_OK_OPTION)
+    } catch (error) {
+      logger.warn(`Could not set advertising "${field}":`, error)
+    }
+  }
+}
+
+// ---------------------------------------------------------------------------
+// Public API — Happy path
+// ---------------------------------------------------------------------------
+
+export async function fillSportsHallReportForm(
+  data: SportsHallReportData,
+  leagueCategory: LeagueCategory,
+  language: Language
+): Promise<Uint8Array> {
+  const { pdfDoc, form } = await loadPdfForm(leagueCategory, language)
+  const mapping = getFieldMapping(leagueCategory)
+
+  fillBaseGameInfo(form, data, mapping)
+
+  // Do not flatten: referees download this form to fill remaining fields themselves
+  return pdfDoc.save()
+}
+
+export async function generateAndDownloadSportsHallReport(
+  data: SportsHallReportData,
+  leagueCategory: LeagueCategory,
+  language: Language
+): Promise<void> {
+  const pdfBytes = await fillSportsHallReportForm(data, leagueCategory, language)
+  const filename = buildReportFilename(
+    leagueCategory,
+    language,
+    data.startingDateTime,
+    data.gameNumber
+  )
+  downloadPdf(pdfBytes, filename)
+}
+
 export interface WizardReportOptions {
   data: SportsHallReportData
   leagueCategory: LeagueCategory
@@ -921,18 +427,7 @@ export async function fillSportsHallReportWizard(
     logger.warn(`Could not check "${wizardMapping.allPointsInOrderCheckbox}":`, error)
   }
 
-  // Set advertising "Werbung auf Spielerkleidung" per team
-  const adEntries: Array<{ field: string; hasAd: boolean }> = [
-    { field: wizardMapping.advertisingHomeTeam, hasAd: jerseyAdvertising.homeTeam },
-    { field: wizardMapping.advertisingAwayTeam, hasAd: jerseyAdvertising.awayTeam },
-  ]
-  for (const { field, hasAd } of adEntries) {
-    try {
-      form.getRadioGroup(field).select(hasAd ? RADIO_OK_OPTION : RADIO_NOT_OK_OPTION)
-    } catch (error) {
-      logger.warn(`Could not set advertising "${field}":`, error)
-    }
-  }
+  fillAdvertisingRadios(form, wizardMapping, jerseyAdvertising)
 
   // Embed first referee signature if provided
   if (signatureDataUrl) {
@@ -947,11 +442,6 @@ export async function fillSportsHallReportWizard(
   return pdfDoc.save()
 }
 
-/**
- * Generates a sports hall report with all checkpoints marked as OK and
- * the first referee's signature embedded. Returns the PDF bytes and filename
- * for use with the Web Share API or print dialog.
- */
 export async function generateWizardReportBytes(
   options: WizardReportOptions & { signatureDataUrl: string }
 ): Promise<{ pdfBytes: Uint8Array; filename: string }> {
@@ -966,97 +456,8 @@ export async function generateWizardReportBytes(
 }
 
 // ---------------------------------------------------------------------------
-// Non-conformant report types and fill function
+// Public API — Non-conformant report
 // ---------------------------------------------------------------------------
-
-/** Which sub-items are flagged as not-OK, keyed by section ID then sub-item ID. */
-export type NonConformantSelections = Record<string, Set<string>>
-
-/** Signature data for the non-conformant workflow (up to 4 signers). */
-export interface NonConformantSignatures {
-  firstReferee?: string
-  secondReferee?: string
-  homeTeamCoach?: { name: string; signature: string }
-  awayTeamCoach?: { name: string; signature: string }
-}
-
-/**
- * Set all checklist radio groups: OK or not-OK per sub-item.
- */
-function fillChecklistRadios(
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  form: any,
-  sections: readonly ChecklistSection[],
-  nonConformantSubItems: NonConformantSelections
-): void {
-  for (const section of sections) {
-    const flaggedItems = nonConformantSubItems[section.id]
-    for (const subItem of section.subItems) {
-      const isNotOk = flaggedItems?.has(subItem.id) ?? false
-      const option = isNotOk ? RADIO_NOT_OK_OPTION : RADIO_OK_OPTION
-      try {
-        form.getRadioGroup(subItem.radioField).select(option)
-      } catch (error) {
-        logger.warn(`Could not set radio "${subItem.radioField}" to ${option}:`, error)
-      }
-    }
-  }
-}
-
-/** Font size for comment fields so longer text fits in the narrow PDF cells */
-const COMMENT_FIELD_FONT_SIZE = 6
-
-/**
- * Fill per-section comments into their respective PDF comment fields.
- */
-function fillNonConformantComments(
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  form: any,
-  sections: readonly ChecklistSection[],
-  nonConformantSubItems: NonConformantSelections,
-  sectionComments: Record<string, string>
-): void {
-  for (const section of sections) {
-    if (!nonConformantSubItems[section.id]?.size) continue
-    const comment = sectionComments[section.id]?.trim()
-    if (!comment) continue
-    const commentField = section.commentFields[0]
-    if (!commentField) continue
-    try {
-      const field = form.getTextField(commentField)
-      field.setFontSize(COMMENT_FIELD_FONT_SIZE)
-      field.setText(comment)
-    } catch (error) {
-      logger.warn(`Could not set comment field "${commentField}":`, error)
-    }
-  }
-}
-
-/**
- * Embed up to 4 signatures at their designated positions.
- */
-async function embedAllSignatures(
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  pdfDoc: any,
-  positions: AllSignaturePositions,
-  signatures: NonConformantSignatures
-): Promise<void> {
-  const entries: Array<[string | undefined, SignaturePosition]> = [
-    [signatures.firstReferee, positions.firstReferee],
-    [signatures.secondReferee, positions.secondReferee],
-    [signatures.homeTeamCoach?.signature, positions.homeTeam],
-    [signatures.awayTeamCoach?.signature, positions.awayTeam],
-  ]
-
-  for (const [dataUrl, position] of entries) {
-    if (!dataUrl) continue
-    try {
-      await embedSignatureAtPosition(pdfDoc, dataUrl, position)
-    } catch (error) {
-      logger.warn('Could not embed signature:', error)
-    }
-  }
-}
 
 export interface NonConformantReportOptions {
   data: SportsHallReportData
@@ -1068,9 +469,6 @@ export interface NonConformantReportOptions {
   signatures?: NonConformantSignatures
 }
 
-/**
- * Fill the sports hall report for the non-conformant workflow.
- */
 export async function fillNonConformantReport(
   options: NonConformantReportOptions & { flatten?: boolean }
 ): Promise<Uint8Array> {
@@ -1086,26 +484,13 @@ export async function fillNonConformantReport(
   const { pdfDoc, form } = await loadPdfForm(leagueCategory, language)
   const mapping = getFieldMapping(leagueCategory)
   const wizardMapping = getWizardFieldMapping(leagueCategory)
-  const sections = getChecklistSections(leagueCategory)
-  const signatureNameFields =
-    leagueCategory === 'NLA' ? NLA_SIGNATURE_NAME_FIELDS : NLB_SIGNATURE_NAME_FIELDS
+  const sections = getPdfChecklistSections(leagueCategory)
+  const signatureNameFields = getSignatureNameFields(leagueCategory)
 
   fillBaseGameInfo(form, data, mapping)
   fillChecklistRadios(form, sections, nonConformantSubItems)
   fillNonConformantComments(form, sections, nonConformantSubItems, sectionComments)
-
-  // Set advertising radio fields (overrides checklist values for these specific fields)
-  const adEntries: Array<{ field: string; hasAd: boolean }> = [
-    { field: wizardMapping.advertisingHomeTeam, hasAd: jerseyAdvertising.homeTeam },
-    { field: wizardMapping.advertisingAwayTeam, hasAd: jerseyAdvertising.awayTeam },
-  ]
-  for (const { field, hasAd } of adEntries) {
-    try {
-      form.getRadioGroup(field).select(hasAd ? RADIO_OK_OPTION : RADIO_NOT_OK_OPTION)
-    } catch (error) {
-      logger.warn(`Could not set advertising "${field}":`, error)
-    }
-  }
+  fillAdvertisingRadios(form, wizardMapping, jerseyAdvertising)
 
   if (signatures?.homeTeamCoach?.name) {
     trySetTextField(form, signatureNameFields.homeTeam, signatures.homeTeamCoach.name)
@@ -1124,9 +509,6 @@ export async function fillNonConformantReport(
   return pdfDoc.save()
 }
 
-/**
- * Generate a non-conformant report preview (no signatures) for review.
- */
 export async function generateNonConformantPreviewBytes(
   options: Omit<NonConformantReportOptions, 'signatures'>
 ): Promise<{ pdfBytes: Uint8Array; filename: string }> {
@@ -1140,10 +522,6 @@ export async function generateNonConformantPreviewBytes(
   return { pdfBytes, filename }
 }
 
-/**
- * Generate the final non-conformant report with all signatures embedded.
- * The form is flattened for download to ensure uniform text rendering.
- */
 export async function generateNonConformantReportBytes(
   options: NonConformantReportOptions & { signatures: NonConformantSignatures }
 ): Promise<{ pdfBytes: Uint8Array; filename: string }> {
